[GridNG] Allow stretching in block-axis.
This block-stretching logic as:
https://chromium-review.googlesource.com/c/chromium/src/+/2513771
... did for the inline-axis.

Previously we didn't have the concept of "stretching" the block-axis
instead relying on the parent layout algorithm to set a fixed
block-size.

This allow stretching (similar to the inline-dimension).

The failing tests (which I'll update in the test expectations) are due
to this specification issue:
https://github.com/w3c/csswg-drafts/issues/5713

The changes in layout_replaced.cc are needed to transfer the stretched
size through the aspect-ratio.

Bug: 1045599
Change-Id: Ifdd8653d81a1ead84a0afbf5ebd31ba0b75dc8ee